Unit tests fragmentation for easy handling. Added fix for empty set and part of operator
