+++ src/engine/__tests__/engine/murmur3/node.spec.js
@@ -22,8 +22,8 @@ const rl = require('readline');
 const utils = require('../../../engine/murmur3');
 
 [
-  'murmur3-sample-data-v2.csv',
-  'murmur3-sample-data-non-alpha-numeric-v2.csv'
+  // 'murmur3-sample-data-v2.csv',
+  // 'murmur3-sample-data-non-alpha-numeric-v2.csv'
 ].forEach(filename => {
 
   tape('MURMUR3 / validate hashing behavior using sample data', assert => {
