+++ src/engine/matchers/part_of.js
@@ -18,18 +18,25 @@ limitations under the License.
 /*eslint-disable eqeqeq */
 
 const log = require('debug')('splitio-engine:matcher');
-const _intersection = require('lodash/intersection');
-const _uniq = require('lodash/uniq');
+const intersection = require('lodash/intersection');
+const uniq = require('lodash/uniq');
 
 function partOfMatcherContext(vo /*: whitelistObject */) /*: Function */ {
   return function partOfMatcher(value /*: array */) /*: boolean */ {
-    let normalizedValue = _uniq(value.map(e => e + ''));
-    // If the intersection returns all of value elements, it is a part of vo.value
-    let partOf = _intersection(normalizedValue, vo.whitelist).length === normalizedValue.length;
+    // Set operations are based on string comparisons
+    const normalizedValue = uniq(value.map(e => e + ''));
+    let isPartOf = false;
 
-    log(`[partOfMatcher] ${value} is part of ${vo.whitelist}? ${partOf}`);
+    if (normalizedValue.length > 0) {
+      // If the intersection returns all of value elements, it is a part of vo.value
+      isPartOf = intersection(normalizedValue, vo.whitelist).length === normalizedValue.length;
 
-    return partOf;
+      log(`[partOfMatcher] ${value} is part of ${vo.whitelist}? ${isPartOf}`);
+    } else {
+      log(`[partOfMatcher] empty set doesn't match`);
+    }
+
+    return isPartOf;
   };
 }
